feat: 커뮤니티 페이지에 이미지 컨트롤러 부착
