+++ salgulok/src/pages/community/WritePage.tsx
