+++ salgulok/src/types/post.ts
